requirements.txt
aws-cdk-lib==2.41.0
constructs==10.0.0,<11.0.0
aws-cdk.aws-apigatewayv2-alpha
aws-cdk.aws-apigatewayv2-integrations-alpha
aws-cdk.aws-apigatewayv2-authorizers-alpha
python-dotenv
boto3
